requirements.txt
pymdown-extensions
Pygments
#Jinja2<3.0.0
Jinja2
markupsafe<2
#markupsafe
mkdocs>=1.1
#mkdocs
#mkdocs-rtd-dropdown
mkdocs-windmill
mkdocs-macros-plugin
pyyaml
